File upload download

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator/check');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -21,9 +22,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => { // only will receive post requests 
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
+
+    const imageUrl = image.path;
+
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -34,7 +53,6 @@ exports.postAddProduct = (req, res, next) => { // only will receive post request
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -72,7 +90,6 @@ exports.postAddProduct = (req, res, next) => { // only will receive post request
             //     validationErrors: []
             // });
             //res.redirect('/500');
-
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error); // passing a error in next function will activated the error handling middleware from express js
@@ -111,8 +128,9 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const updatedImage = req.file;
     const updatedDesc = req.body.description;
+
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
@@ -123,7 +141,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
                 _id: prodId
@@ -141,7 +158,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.imageUrl = updatedImageUrl;
+            if (updatedImage) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = updatedImage.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('UPDATED PRODUCT!');
@@ -178,7 +198,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
